libarchive: Always ensure we have a label
The way SELinux works, there isn't a concept of "something without
a label".  There *is* the `unlabeled_t` type that the kernel makes
up when it finds no label on a mounted filesystem.  But that's still
a label, it's just a kernel fallback.

When we go to import a tar archive, in most cases it won't
have SELinux labels, and even if it does, we want to overwrite it.

Now, the concept of a "whiteout" file arose relatively recently
as a special file type that's used in the container ecosystem alongside
overlayfs.

For some reason, presumably due to its special nature, at least the
Fedora SELinux policy doesn't define a label for it.

However, when it lands on disk, the kernel will still assign it
a default label.

And that's the problem - the computed checksum won't match, which
will appear as corruption.

One doesn't see this problem when e.g. extracting a tarball
temporarily to disk, and then committing it because the act of extracting
to disk will assign a label.

This problem was hit when importing container layers for ostree-container
work in ostree-rs-ext.