sysroot: Add auto cleanup API
This is a wrapper around `ostree_sysroot_cleanup` that only performs the
cleanup when the `.cleanup` file exists. When the cleanup has completed,
the file is deleted. The purpose of this API is to allow other routines
to indicate that cleanup is needed when they're not in a good position
to initiate the cleanup themselves. This API can then be run at a more
opportune time without actually initiating the expensive cleanup
operations unless they were needed.