deploy: Touch auto cleanup file when finalizing staged deployment
Currently finalizing a staged deployment skips the pruning that
typically happens when writing out a deployment since it would block the
shutdown path. In order to leave an indication that a cleanup should be
run at a more opportune time, touch the `.cleanup` file after the staged
deployment is written out. The `ostree_sysroot_auto_cleanup` API can
then be used to perform the pruning at a better time without fear that
it will initiate an unnecessary prune.