boot: Add systemd unit to cleanup sysroot automatically
This unit makes use of the `--auto` option of `ostree admin cleanup` to
cleanup the sysroot when the `/sysroot/.cleanup` file is left behind
when finalizing a staged deployment. The purpose of this service is to
restore pruning of deleted deployments when a staged deployment is
written out during shutdown of the previous boot. The unit is optional
as the cleanup can be run at another time or not at all.

Cleaning the sysroot will prune the repo, and this can be a slow and IO
intensive operation. To keep the system from blocking, the default
`Before` dependency on `multi-user.target` has been removed. Since the
service will then run in the background, the IO scheduling class has
been lowered to `idle` to keep the system's primary tasks more
responsive.