lib/repo: Support `/etc/ostree/config.d/`
The sysroot handling of libostree supports multiple knobs specified in
the system repo keyfile. E.g. `sysroot.bootloader`, `sysroot.readonly`,
etc...

One issue is that all the knobs so far have been "vendor-oriented". I.e.
they're not necessarily something we expect users to directly modify.

I'd like to add more sysroot configurable knobs which we *do* expect
users to modify, but I don't want them to modify the main config
directly and mix with knobs they really shouldn't touch. It's also not
convenient to modify in a provisioning flow.

Add a new concept of a "user config" which is a config obtained by
merging all the configs in `/etc/ostree/config.d`. Now, there's no
official API for merging `GKeyFile` objects, but the parser has the
right semantics if we simply concatenate all the configs and parse the
resulting final config: keys are merged into the same group and later
keys override earlier keys. Appending to a string list is not supported,
but we can design our knobs so that this is not an issue.

We don't want user configs to be able to override e.g. `core.mode`. We
only want to support specific keys in those configs (but for
consistency, also in the primary config). So keep it as a separate
`user_config` object.